+++ libraries/resource_chef_dk_app_debian.rb
@@ -1,10 +1,11 @@
 # encoding: utf-8
 # frozen_string_literal: true
+
 #
 # Cookbook Name:: chef-dk
 # Library:: resource_chef_dk_app_debian
 #
-# Copyright 2014-2016, Jonathan Hartman
+# Copyright 2014-2017, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -103,7 +104,7 @@ class Chef
         # Chef-DK, so the remove action is just a package purge regardless of
         # installation type.
         #
-        %w(remove_direct! remove_repo! remove_custom!).each do |m|
+        %w[remove_direct! remove_repo! remove_custom!].each do |m|
           define_method(m) { package('chefdk') { action :purge } }
         end
       end