+++ metadata.rb
@@ -4,7 +4,7 @@
 name 'chef-dk'
 maintainer 'Jonathan Hartman'
 maintainer_email 'j@p4nt5.com'
-license 'Apache v2.0'
+license 'Apache-2.0'
 description 'Installs/configures the Chef-DK'
 long_description 'Installs/configures the Chef-DK'
 version '3.1.1'
@@ -13,17 +13,17 @@ chef_version '>= 12'
 source_url 'https://github.com/roboticcheese/chef-dk-chef'
 issues_url 'https://github.com/roboticcheese/chef-dk-chef/issues'
 
-depends 'apt-chef', '~> 2.0'
-depends 'yum-chef', '~> 2.0'
-depends 'homebrew', '~> 2.1'
-depends 'dmg', '~> 3.0'
-depends 'chocolatey', '~> 1.0'
+depends 'apt-chef', '< 3.0'
+depends 'yum-chef', '< 4.0'
+depends 'homebrew', '< 5.0'
+depends 'dmg', '< 4'
+depends 'chocolatey', '< 2'
 
-supports 'ubuntu', '>= 12.04'
-supports 'debian', '>= 6.0'
-%w(redhat centos scientific amazon).each do |os|
-  supports       os, '>= 6.0'
+supports 'ubuntu'
+supports 'debian'
+%w[redhat centos scientific amazon].each do |os|
+  supports os
 end
 supports 'fedora'
-supports 'mac_os_x', '>= 10.8'
+supports 'mac_os_x'
 supports 'windows'