+++ spec/resources/chef_dk/mac_os_x/10_12_spec.rb
@@ -3,11 +3,11 @@
 
 require_relative '../../chef_dk'
 
-describe 'resources::chef_dk::mac_os_x::10_10' do
+describe 'resources::chef_dk::mac_os_x::10_12' do
   include_context 'resources::chef_dk'
 
   let(:platform) { 'mac_os_x' }
-  let(:platform_version) { '10.10' }
+  let(:platform_version) { '10.12' }
 
   it_behaves_like 'any platform'
 end