+++ spec/resources/chef_dk_app/mac_os_x/10_12_spec.rb
@@ -3,11 +3,11 @@
 
 require_relative '../mac_os_x'
 
-describe 'resources::chef_dk_app::mac_os_x::10_10' do
+describe 'resources::chef_dk_app::mac_os_x::10_12' do
   include_context 'resources::chef_dk_app::mac_os_x'
 
   let(:platform) { 'mac_os_x' }
-  let(:platform_version) { '10.10' }
+  let(:platform_version) { '10.12' }
 
-  it_behaves_like 'any Mac OS X platform'
+  it_behaves_like 'any MacOS platform'
 end