+++ test/fixtures/cookbooks/chef-dk_test/attributes/default.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 unless node['platform_family'] == 'windows'
-  default['chef_dk']['shell_users'] = %w(root)
+  default['chef_dk']['shell_users'] = %w[root]
 end
 
-default['chef_dk']['gems'] = %w(cabin)
+default['chef_dk']['gems'] = %w[cabin]